Fold site config and probe targets into the sightmap package
Three tooling helpers that lived in the CLI adapter layer move down into
the sightmap library, where external callers (and the library's own
Corpus) can reuse them:

- .sightmap/config.yaml parsing: SiteConfig + loadSiteConfig become
  sightmap.SiteConfig + sightmap.LoadConfig. It is tooling-only config
  (not spec data), but the file lives inside .sightmap/, so the package
  owns its parse.
- whole-corpus probe URLs: corpusProbeTargets + probeTarget become
  Corpus.ProbeTargets + sightmap.ProbeTarget, reusing View.SnapBasename
  (which drops the duplicate sanitizeBasename helper). The capture --all
  and sel-probe --all paths now load the corpus once instead of twice.
- collapse sightmap.DirLoader(dir).Load() to sightmap.Load(dir) at the
  lint/validate/serve/report call sites.

No behavior change; deletes cmd_probe_urls.go and siteconfig.go.
