+++ go/cmd/sightmap/cmd_validate.go
@@ -15,8 +15,7 @@ func runValidate(args []string) error {
 		return err
 	}
 
-	loader := sightmap.DirLoader(*sightmapDir)
-	corpus, err := loader.Load()
+	corpus, err := sightmap.Load(*sightmapDir)
 	if err != nil {
 		return fmt.Errorf("load corpus: %w", err)
 	}
